small change to installation from source scripts
git-svn-id: file:///home/karl/SVN2/amilab.svn/code/trunk@1006 52b0d089-85ac-4fb5-acf5-4e3172ea7aaf

+++ install_fedora13.sh
@@ -1,4 +1,22 @@
-#!/bin/sh
+#!/bin/bash
+
+# check for proper number of command line arguments
+ERROR_PAR=65
+
+if [ $# -ne 1 ]
+   then
+      echo
+      echo " Usage: "
+      echo " $0 amilabtag"
+      echo " where amilabtag is the name of the amilab subversion repository tab to download (for example release-3.0.0)"
+      echo
+      exit $ERROR_PAR
+fi
+
+#
+# need to specify the release tag as argument
+#
+releasetag=$1
 
 #
 # Script for installing amilab from source on Ubuntu, tested on Ubuntu 10.10 32 bits
@@ -26,17 +44,20 @@ yum install gawk
 numthreads=`awk '/model name/  {ORS=""; count++;  }  END {  print  count "\n" }' /proc/cpuinfo`
 
 # compilation of ITK
-yum install wget tar
-wget http://voxel.dl.sourceforge.net/sourceforge/itk/InsightToolkit-3.20.0.tar.gz
-tar zxf InsightToolkit-3.20.0.tar.gz
-cd InsightToolkit-3.20.0
-mkdir build_release
-cd build_release
-cmake  -DBUILD_EXAMPLES=OFF -DBUILD_TESTING=OFF -DCMAKE_BUILD_TYPE=Release ..
-make -j ${numthreads} install
-cd ../..
+if [ \! -f InsightToolkit-3.20.0.tar.gz ];
+then
+  yum install wget tar
+  wget http://voxel.dl.sourceforge.net/sourceforge/itk/InsightToolkit-3.20.0.tar.gz
+  tar zxf InsightToolkit-3.20.0.tar.gz
+  cd InsightToolkit-3.20.0
+  mkdir build_release
+  cd build_release
+  cmake  -DBUILD_EXAMPLES=OFF -DBUILD_TESTING=OFF -DCMAKE_BUILD_TYPE=Release ..
+  make -j ${numthreads} install
+  cd ../..
+fi
 
-svn co  https://amilab.svn.sourceforge.net/svnroot/amilab/tags/stable amilab_stable
+svn co  https://amilab.svn.sourceforge.net/svnroot/amilab/tags/${releasetag} amilab_stable
 cd amilab_stable
 maindir=`pwd`
